requirements.txt
mysql-replication==0.27
loguru==0.5.3
PySide6==6.2.2.1
